// optika-admin-ui/src/ts-types/api.enums.ts
/* tslint:disable */
// Generated using typescript-generator version 2.12.476 on 2022-07-21 20:41:05.

export enum Category {
    MEN = "MEN",
    WOMEN = "WOMEN",
    UNISEX = "UNISEX",
    KIDS = "KIDS",
}

export enum CategoryEx {
    MEN = "MEN",
    WOMEN = "WOMEN",
    UNISEX = "UNISEX",
    KIDS = "KIDS",
    ALL = "ALL",
}

export enum ProductOrderBy {
    PRICE_ASCENDING = "PRICE_ASCENDING",
    PRICE_DESCENDING = "PRICE_DESCENDING",
}

export enum ProductType {
    FRAMES = "FRAMES",
    SUNGLASSES = "SUNGLASSES",
    CONTACT_LENSES = "CONTACT_LENSES",
    CLEAN_AND_MAINTENANCE = "CLEAN_AND_MAINTENANCE",
    OTHER = "OTHER",
}

export enum ProductTypeEx {
    FRAMES = "FRAMES",
    SUNGLASSES = "SUNGLASSES",
    CONTACT_LENSES = "CONTACT_LENSES",
    CLEAN_AND_MAINTENANCE = "CLEAN_AND_MAINTENANCE",
    OTHER = "OTHER",
    ALL = "ALL",
}